ESSnapshotsManager class created
